PaSST/requirements.txt
av>=10.0.0
h5py>=3.8.0
jsonpickle>=3.0.1
kk-sacred>=0.8.4
librosa>=0.10.0.post2
timm>=0.4.12
torchmetrics>=0.2.0
pytorch-lightning<2.0.0
wandb>=0.14.2
tqdm
copy
matplotlib
pandas
numpy
sklearn
ast
collections